clean up Swift implementation
